requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pandas
numpy
scikit-learn
python-dotenv
jinja2
httpx
